Flare cooldown upgrade now works. Fixed visual bug with the flare upgrade. Added check for if the player has upgraded an upgrade to the maximum level (changes button to gold).

+++ Wingsuiter Game/Assets/Scripts/Managers/GUITextManager.cs
@@ -12,7 +12,7 @@ public class GUITextManager : MonoBehaviour
     public TMP_Text speedText;
     public TMP_Text coinText;
 
-    public void UpdateFlareText(int time, int delay)
+    public void UpdateFlareText(int time, float delay)
     {
         // Calculate the green value
         float greenValue = (float) (delay - time) / delay;
